Fix analysis issues
