+++ MediaBrowser.Providers/Plugins/AudioDb/AudioDbAlbumImageProvider.cs
