+++ MediaBrowser.Providers/Plugins/AudioDb/AudioDbArtistImageProvider.cs
