+++ MediaBrowser.Providers/Plugins/Omdb/OmdbProvider.cs
@@ -1,4 +1,4 @@
-#pragma warning disable CS1591
+#pragma warning disable CS159, SA1300
 
 using System;
 using System.Collections.Generic;
@@ -20,6 +20,7 @@ using MediaBrowser.Model.IO;
 
 namespace MediaBrowser.Providers.Plugins.Omdb
 {
+    /// <summary>Provider for OMDB service.</summary>
     public class OmdbProvider
     {
         private readonly IFileSystem _fileSystem;
@@ -29,6 +30,11 @@ namespace MediaBrowser.Providers.Plugins.Omdb
         private readonly IApplicationHost _appHost;
         private readonly JsonSerializerOptions _jsonOptions;
 
+        /// <summary>Initializes a new instance of the <see cref="OmdbProvider"/> class.</summary>
+        /// <param name="httpClientFactory">HttpClientFactory to use for calls to OMDB service.</param>
+        /// <param name="fileSystem">IFileSystem to use for store OMDB data.</param>
+        /// <param name="appHost">IApplicationHost to use.</param>
+        /// <param name="configurationManager">IServerConfigurationManager to use.</param>
         public OmdbProvider(IHttpClientFactory httpClientFactory, IFileSystem fileSystem, IApplicationHost appHost, IServerConfigurationManager configurationManager)
         {
             _httpClientFactory = httpClientFactory;
@@ -41,6 +47,14 @@ namespace MediaBrowser.Providers.Plugins.Omdb
             _jsonOptions.Converters.Add(new JsonOmdbNotAvailableInt32Converter());
         }
 
+        /// <summary>Fetches data from OMDB service.</summary>
+        /// <param name="itemResult">Metadata about media item.</param>
+        /// <param name="imdbId">IMDB ID for media.</param>
+        /// <param name="language">Media language.</param>
+        /// <param name="country">Country of origin.</param>
+        /// <param name="cancellationToken">CancellationToken to use for operation.</param>
+        /// <typeparam name="T">The first generic type parameter.</typeparam>
+        /// <returns>Returns a Task object that can be awaited.</returns>
         public async Task Fetch<T>(MetadataResult<T> itemResult, string imdbId, string language, string country, CancellationToken cancellationToken)
             where T : BaseItem
         {
@@ -105,6 +119,17 @@ namespace MediaBrowser.Providers.Plugins.Omdb
             ParseAdditionalMetadata(itemResult, result);
         }
 
+        /// <summary>Gets data about an episode.</summary>
+        /// <param name="itemResult">Metadata about episode.</param>
+        /// <param name="episodeNumber">Episode number.</param>
+        /// <param name="seasonNumber">Season number.</param>
+        /// <param name="episodeImdbId">Episode ID.</param>
+        /// <param name="seriesImdbId">Season ID.</param>
+        /// <param name="language">Episode language.</param>
+        /// <param name="country">Country of origin.</param>
+        /// <param name="cancellationToken">CancellationToken to use for operation.</param>
+        /// <typeparam name="T">The first generic type parameter.</typeparam>
+        /// <returns>Whether operation was successful.</returns>
         public async Task<bool> FetchEpisodeData<T>(MetadataResult<T> itemResult, int episodeNumber, int seasonNumber, string episodeImdbId, string seriesImdbId, string language, string country, CancellationToken cancellationToken)
             where T : BaseItem
         {
@@ -236,6 +261,9 @@ namespace MediaBrowser.Providers.Plugins.Omdb
             return false;
         }
 
+        /// <summary>Gets OMDB URL.</summary>
+        /// <param name="query">Appends query string to URL.</param>
+        /// <returns>OMDB URL with optional query string.</returns>
         public static string GetOmdbUrl(string query)
         {
             const string Url = "https://www.omdbapi.com?apikey=2c9d9507";
@@ -327,6 +355,12 @@ namespace MediaBrowser.Providers.Plugins.Omdb
             return path;
         }
 
+        /// <summary>Gets response from OMDB service as type T.</summary>
+        /// <param name="httpClient">HttpClient instance to use for service call.</param>
+        /// <param name="url">Http URL to use for service call.</param>
+        /// <param name="cancellationToken">CancellationToken to use for service call.</param>
+        /// <typeparam name="T">The first generic type parameter.</typeparam>
+        /// <returns>OMDB service response as type T.</returns>
         public async Task<T> GetDeserializedOmdbResponse<T>(HttpClient httpClient, string url, CancellationToken cancellationToken)
         {
             using var response = await GetOmdbResponse(httpClient, url, cancellationToken).ConfigureAwait(false);
@@ -335,6 +369,11 @@ namespace MediaBrowser.Providers.Plugins.Omdb
             return await JsonSerializer.DeserializeAsync<T>(content, _jsonOptions, cancellationToken).ConfigureAwait(false);
         }
 
+        /// <summary>Gets response from OMDB service.</summary>
+        /// <param name="httpClient">HttpClient instance to use for service call.</param>
+        /// <param name="url">Http URL to use for service call.</param>
+        /// <param name="cancellationToken">CancellationToken to use for service call.</param>
+        /// <returns>OMDB service response as HttpResponseMessage.</returns>
         public static Task<HttpResponseMessage> GetOmdbResponse(HttpClient httpClient, string url, CancellationToken cancellationToken)
         {
             return httpClient.GetAsync(url, cancellationToken);
@@ -538,10 +577,13 @@ namespace MediaBrowser.Providers.Plugins.Omdb
             }
         }
 
+        /// <summary>Describes OMDB rating.</summary>
         public class OmdbRating
         {
+            /// <summary>Gets or sets rating source.</summary>
             public string Source { get; set; }
 
+            /// <summary>Gets or sets rating value.</summary>
             public string Value { get; set; }
         }
     }
